fastcnn/requirements.txt
azure-functions
onnxruntime
numpy
Pillow
onnx
matplotlib
line-profiler